Fixes to build

+++ src/KAST.Core/Interfaces/IServerConfigService.cs
@@ -47,7 +47,7 @@ public class BasicConfigData
     public int MaxSizeGuaranteed { get; set; } = 512;
     public int MaxSizeNonguaranteed { get; set; } = 256;
     public int MinBandwidth { get; set; } = 131072;
-    public int MaxBandwidth { get; set; } = 10000000000;
+    public long MaxBandwidth { get; set; } = 10000000000;
     public double MinErrorToSend { get; set; } = 0.001;
     public double MinErrorToSendNear { get; set; } = 0.01;
     public int MaxCustomFileSize { get; set; }
